requirements.txt
flask==2.0.1
h5py==2.10.0
Markdown==3.2.2
MarkupSafe==2.0
numpy==1.18.0
oauthlib==3.1.0
pillow==7.2.0
protobuf==3.12.4
PyYAML==5.4.1
python-dotenv==0.14.0
requests==2.25.1
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorflow-cpu==2.3.0
urllib3==1.26.5
line-bot-sdk==1.16.0
openai==1.35.10
python-dotenv
tqdm
langchain_community
langchain-core
langchain-openai
langchain
